Add Mighty Mouse support from Dirk Engling.
Patch fetched from: https://reviews.freebsd.org/D3702

I'm not convinced that the FreeBSD BT stack's architecture is ideal;
especially in terms of having devices fail to enumerate new features
without first caching their HID details in a "config file" that's
essentially machine-generated.

This works for now, but bluetooth is a hassle in general. The NetBSD
stack seems to have a cleaner design (and already supports the Magic
Mouse); it may be worth investigating replacement of the FreeBSD
stack with the NetBSD one at a later date.